% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{accuracy_arima}
\alias{accuracy_arima}
\title{Calculate accuracy measue based on ARIMA models}
\usage{
accuracy_arima(ts_info, function_name, length_out)
}
\arguments{
\item{ts_info}{list containing training and test part of a time series}

\item{function_name}{function to calculate the accuracy function, the arguments of this function
should be forecast, training and test set of the time series}

\item{length_out}{number of measures calculated by the function}
}
\value{
a list which contains the accuracy and name of the specific ARIMA model.
}
\description{
Calculate accuracy measue based on ARIMA models
}
