% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{accuracy_functions}
\alias{accuracy_functions}
\alias{accuracy_ets}
\alias{accuracy_arima}
\alias{accuracy_rw}
\alias{accuracy_rwd}
\alias{accuracy_wn}
\alias{accuracy_theta}
\alias{accuracy_stlar}
\alias{accuracy_nn}
\alias{accuracy_snaive}
\alias{accuracy_mstl}
\alias{accuracy_tbats}
\title{Calculate accuracy measure from different forecasting methods}
\usage{
accuracy_ets(ts_info, function_name)

accuracy_arima(ts_info, function_name)

accuracy_rw(ts_info, function_name)

accuracy_rwd(ts_info, function_name)

accuracy_wn(ts_info, function_name)

accuracy_theta(ts_info, function_name)

accuracy_stlar(ts_info, function_name)

accuracy_nn(ts_info, function_name)

accuracy_snaive(ts_info, function_name)

accuracy_mstl(ts_info, function_name)

accuracy_tbats(ts_info, function_name)
}
\arguments{
\item{ts_info}{list containing training and test part of a time series}

\item{function_name}{function to calculate the accuracy function, the arguments of this function
should be forecast, training and test set of the time series}
}
\value{
\code{accuracy_ets} returns a list which contains the accuracy and name of the specific ETS model.
\code{accuracy_arima} a list which contains the accuracy and name of the specific ARIMA model.
\code{accuracy_rw} returns accuracy measure calculated baded on random walk model
\code{accuracy_rwd} returns accuracy measure calculated baded on random walk with drift model
\code{accuracy_wn} returns accuracy measure calculated based on white noise model
\code{accuracy_theta} returns accuracy measure calculated based on theta method
\code{accuracy_stlar} returns accuracy measure calculated based on stlar method
\code{accuracy_nn} returns accuracy measure calculated based on neural network forecasts
\code{accuracy_snaive} returns accuracy measure calculated based on snaive method
\code{accuracy_mstl} returns accuracy measure calculated based on multiple seasonal decomposition
\code{accuracy_tbats} returns accuracy measure calculated based on TBATS models
}
\description{
Calculate accuracy measure from different forecasting methods
}
