% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{accuracy_mstl}
\alias{accuracy_mstl}
\title{Calculate accuracy based on MSTL}
\usage{
accuracy_mstl(ts_info, function_name, length_out)
}
\arguments{
\item{ts_info}{list containing training and test part of a time series}

\item{function_name}{function to calculate the accuracy function, the arguments of this function
should be forecast, training and test set of the time series}

\item{length_out}{number of measures calculated by the function}
}
\value{
accuracy measure calculated based on multiple seasonal decomposition
}
\description{
Calculate accuracy based on MSTL
}
