% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{accuracy_nn}
\alias{accuracy_nn}
\title{Calculate accuracy measure calculated based on neural network forecasts}
\usage{
accuracy_nn(ts_info, function_name)
}
\arguments{
\item{ts_info}{list containing training and test part of a time series}

\item{function_name}{function to calculate the accuracy function, the arguments of this function
should be forecast, training and test set of the time series}
}
\value{
accuracy measure calculated based on neural network forecasts
}
\description{
Calculate accuracy measure calculated based on neural network forecasts
}
