% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{accuracy_rwd}
\alias{accuracy_rwd}
\title{Calculate accuracy measure based on random walk with drift}
\usage{
accuracy_rwd(ts_info, function_name)
}
\arguments{
\item{ts_info}{list containing training and test part of a time series}

\item{function_name}{function to calculate the accuracy function, the arguments of this function
should be forecast, training and test set of the time series}
}
\value{
accuracy measure calculated baded on random walk with drift model
}
\description{
Calculate accuracy measure based on random walk with drift
}
