% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{accuracy_wn}
\alias{accuracy_wn}
\title{Calculate accuracy measure based on white noise process}
\usage{
accuracy_wn(ts_info, function_name, length_out)
}
\arguments{
\item{ts_info}{list containing training and test part of a time series}

\item{function_name}{function to calculate the accuracy function, the arguments of this function
should be forecast, training and test set of the time series}

\item{length_out}{number of measures calculated by the function}
}
\value{
returns accuracy measure calculated based on white noise process
}
\description{
Calculate accuracy measure based on white noise process
}
