% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{acf_seasonalDiff}
\alias{acf_seasonalDiff}
\title{Autocorrelation coefficients based on seasonally differenced series}
\usage{
acf_seasonalDiff(y, m, lagmax)
}
\arguments{
\item{y}{a univariate time series}

\item{m}{frequency of the time series}

\item{lagmax}{maximum lag at which to calculate the acf}
}
\value{
A vector of 3 values: first ACF value of seasonally-differenced series, ACF value at the first seasonal lag of seasonally-differenced series,
sum of squares of first 5 autocorrelation coefficients of seasonally-differenced series.
}
\description{
Autocorrelation coefficients based on seasonally differenced series
}
\author{
Thiyanga Talagala
}
