% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{add_exit_rule}
\alias{add_exit_rule}
\title{Agrega una regla de salida del mercado}
\usage{
add_exit_rule(
  strategy,
  fees,
  sigcol = "exitLong",
  label = paste0("rule_", sigcol),
  orderside = "long"
)
}
\arguments{
\item{strategy}{el objeto strategy}

\item{label}{nombre a utilizar para la regla}

\item{orderside}{tipo de operacion (long | short)}

\item{sigCol}{nombre del signal de salida}
}
\value{
la estrategia con esta regla agregada
}
\description{
Agrega una regla de salida del mercado
}
