% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{apply_indicator_signals}
\alias{apply_indicator_signals}
\title{Aplica los indicadores y señales al mktdata}
\usage{
apply_indicator_signals(strategy, mktdata)
}
\arguments{
\item{strategy}{objecto strategy}

\item{mktdata}{xts con datos OHLC}
}
\value{
xts con las columnas adicionales de indicadores y señales
}
\description{
Aplica los indicadores y señales al mktdata
}
