% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest.opt}
\alias{backtest.opt}
\title{Realiza el backtest sin hacer graficos.}
\usage{
backtest.opt(
  strategy,
  symbols,
  geometric = FALSE,
  initial_equity = 10000,
  silent = FALSE
)
}
\arguments{
\item{strategy}{el strategy object}

\item{symbols}{vector de nombres de datasets a utilizar en el backtest. Ver detalles.}

\item{geometric}{utilizar media geometrica para calcular returns.}

\item{initial_equity}{fondos de la cuenta al empezar el backtest}

\item{silent}{no imprime las transacciones realizadas}
}
\value{
lista con resultados del backtest
}
\description{
Realiza el backtest sin hacer graficos.
}
\details{
Los datasets especificados en `symbols` deben estar cargados en el entorno global
como workaround a las limitaciones de blotter/quantstrat.
}
\examples{
\dontrun{
data("ES35_D1")
backtest.opt(macd(), "ES35_D1")
}
}
