% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbacktest.R
\name{backtest}
\alias{backtest}
\title{Run the backtest with the specified parametrs}
\usage{
backtest(
  data,
  strat,
  cost,
  qty,
  sell_at_end = T,
  price_fun = quantmod::Op,
  debug = F
)
}
\arguments{
\item{data}{xts data}

\item{strat}{strategy list with the strat_fun}

\item{cost}{trade cost}

\item{qty}{order size}

\item{sell_at_end}{sell at the end of the backtest.}

\item{price_fun}{price function for buying and selling. By default, the opening price}

\item{debug}{return additional data for debugging}
}
\value{

}
\description{
Run the backtest with the specified parametrs
}
