% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbacktest.R
\name{backtest_stats}
\alias{backtest_stats}
\title{Calculate backtest stats}
\usage{
backtest_stats(result)
}
\arguments{
\item{result}{results of the backtest}
}
\value{

}
\description{
Calculate backtest stats
}
