% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rf.R
\name{build_rf}
\alias{build_rf}
\title{build random forest classifier}
\usage{
build_rf(
  training_set,
  testset = FALSE,
  rf_type = c("ru", "rcp"),
  ntree,
  seed,
  import = FALSE,
  mtry = 8
)
}
\arguments{
\item{training_set}{data frame of features and class labels}

\item{testset}{features of new time series, default FALSE if a testset is not available}

\item{rf_type}{whether ru(random forest based on unbiased sample) or rcp(random
forest based on class priors)}

\item{ntree}{number of trees in the forest}

\item{seed}{a value for seed}

\item{import}{Should importance of predictors be assessed?, TRUE of FALSE}

\item{mtry}{number of features to be selected at each node}
}
\value{
a list containing the random forest and forecast-models for new series
}
\description{
train a random forest model and predict forecast-models for new series
}
