% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{buy_and_hold}
\alias{buy_and_hold}
\title{Buy and Hold strategy}
\usage{
buy_and_hold(fees = 0, order_size = 100)
}
\arguments{
\item{order_size}{}
}
\value{
un objeto Strategy
}
\description{
Buy and Hold strategy
}
\examples{
strat <- buy_and_hold()
}
