% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_avg.R
\name{cal_avg}
\alias{cal_avg}
\title{Mean of MASE and sMAPE}
\usage{
cal_avg(training, test, forecast)
}
\arguments{
\item{training}{training period of a time series}

\item{test}{test peiod of a time series}

\item{forecast}{forecast obtained from a fitted to the training period}
}
\value{
returns a single value: mean on MASE and sMAPE
}
\description{
Calculate mean of MASE and sMAPE for an individual time series
}
\author{
Thiyanga Talagala
}
