% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_cal.R
\name{cal_features}
\alias{cal_features}
\title{Calculate features for new time series instances}
\usage{
cal_features(ts_data_set, seasonal = FALSE, m = 1, lagmax = 2L)
}
\arguments{
\item{ts_data_set}{a list of univariate time series}

\item{seasonal}{if FALSE, restricts to features suitable for non-seasonal data}

\item{m}{frequency of the time series}

\item{lagmax}{maximum lag at which to calculate the acf (quarterly series-5L and monthly-13L)}
}
\value{
dataframe: each column represent a feature and each row represent a time series
}
\description{
Computes relevant time series features before applying them to the model
}
\author{
Thiyanga Talagala
}
