% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_m4measures.R
\name{cal_m4measures}
\alias{cal_m4measures}
\title{Mean of MASE and sMAPE}
\usage{
cal_m4measures(training, test, forecast)
}
\arguments{
\item{training}{training period of a time series}

\item{test}{test peiod of a time series}

\item{forecast}{forecast obtained from a fitted to the training period}
}
\value{
returns a single value: mean on MASE and sMAPE
}
\description{
Calculate MASE and sMAPE for an individual time series
}
\examples{
require(Mcomp)
require(magrittr)
ts <- Mcomp::M3[[1]]$x
fcast_arima <- auto.arima(ts) \%>\% forecast(h=6)
cal_m4measures(M3[[1]]$x, M3[[1]]$xx, fcast_arima$mean)
}
\author{
Thiyanga Talagala
}
