% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_MASE.R
\name{cal_MASE}
\alias{cal_MASE}
\title{Mean Absolute Scaled Error(MASE)}
\usage{
cal_MASE(training, test, forecast)
}
\arguments{
\item{training}{training peiod of the time series}

\item{test}{test period of the time series}

\item{forecast}{forecast values of the series}
}
\value{
returns a single value
}
\description{
Calculation of mean absolute scaled error
}
\author{
Thiyanga Talagala
}
