% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_medianscaled.R
\name{cal_medianscaled}
\alias{cal_medianscaled}
\title{scale MASE and sMAPE by median}
\usage{
cal_medianscaled(x)
}
\arguments{
\item{x}{output form the function fcast_accuracy, where the parameter accuracyFun = cal_m4measures}
}
\value{
a list with accuracy matrix, vector of arima models and vector of ets models the accuracy
for each forecast-method is average of scaled-MASE and scaled-sMAPE. Median of MASE and sMAPE calculated
based on forecast produced from different models for a given series.
}
\description{
Given a matrix of MASE and sMAPE for each forecasting method and scaled by median and take the mean of
MASE-scaled by median and sMAPE-scaled by median as the forecast accuracy measure to identify the class labels
}
