% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_sMAPE.R
\name{cal_sMAPE}
\alias{cal_sMAPE}
\title{symmetric Mean Absolute Pecentage Error(sMAPE)}
\usage{
cal_sMAPE(training, test, forecast)
}
\arguments{
\item{training}{training peiod of the time series}

\item{test}{test period of the time series}

\item{forecast}{forecast values of the series}
}
\value{
returns a single value
}
\description{
Calculation of symmetric mean absolute percentage error
}
\author{
Thiyanga Talagala
}
