% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_WA.R
\name{cal_WA}
\alias{cal_WA}
\title{Weighted Average}
\usage{
cal_WA(training, test, forecast)
}
\arguments{
\item{training}{training period of a time series}

\item{test}{test peiod of a time series}

\item{forecast}{forecast obtained from a fitted to the training period}
}
\value{
returns a single value: WA based on MASE and sMAPE
}
\description{
Weighted Average(WA) calculated based on MASE, sMAPE for an individual time series
}
\author{
Thiyanga Talagala
}
