% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{checkBlotterUpdate}
\alias{checkBlotterUpdate}
\title{Verifica que no haya inconsistencias entre el portfolio y el account}
\usage{
checkBlotterUpdate(
  port.st = portofolio.name,
  account.st = account.name,
  verbose = TRUE
)
}
\arguments{
\item{port.st}{nombre del portfolio}

\item{account.st}{nombre de la cuenta}

\item{verbose}{muestra mensajes de inconsistencias}
}
\value{
FALSE si hay inconsistencias, TRUE caso contrario
}
\description{
Verifica que no haya inconsistencias entre el portfolio y el account
}
