% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_labels.R
\name{classify_labels}
\alias{classify_labels}
\title{Classify labels according to the FFORMS famework}
\usage{
classify_labels(df_final)
}
\arguments{
\item{df_final}{a dataframe: output from split_names function}
}
\value{
a vector of class labels in FFORMS framewok
}
\description{
This function further classify class labels as in FFORMS framework
}
