% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_msts.R
\name{convert_msts}
\alias{convert_msts}
\title{Convert multiple frequency time series into msts object}
\usage{
convert_msts(y, category)
}
\arguments{
\item{y}{univariate time series}

\item{category}{frequency data have been collected}
}
\value{
a ts object or msts object
}
\description{
Convert multiple frequency(daily, hourly, half-hourly, minutes, seconds)
time series into msts object.
}
