% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbacktest.R
\name{crossover}
\alias{crossover}
\title{Remove repetitions in a vector}
\usage{
crossover(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{

}
\description{
Remove repetitions in a vector
}
\examples{
crossover(c(1, 0, 0, 1, 1, 1, 1, -1, -1))
}
