% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{default_rule_arguments}
\alias{default_rule_arguments}
\title{Retorna una lista con argumentos por defecto}
\usage{
default_rule_arguments(order_size, fees, ...)
}
\arguments{
\item{order_size}{numero o "all"}

\item{...}{argumentos adicionales de ruleSignal. Reemplaza a los defaults}
}
\value{

}
\description{
Retorna una lista con argumentos por defecto
}
\examples{
default_rule_arguments(100, 0, orderside = "long")
}
