% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{df_to_xts}
\alias{df_to_xts}
\title{Convierte un data.frame a un xts}
\usage{
df_to_xts(x, order_by = index)
}
\arguments{
\item{x}{datos a utilizar. Todas las columnas deben ser numericas}

\item{order_by}{columna a utilizar para el indice temporal.}
}
\value{
un xts
}
\description{
Convierte un data.frame a un xts
}
