% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{.sell_at_end}
\alias{.sell_at_end}
\title{Cierra las ordenes pendientes al final del backtest}
\usage{
.sell_at_end(strat, fees)
}
\arguments{
\item{strat}{objeto strategy}

\item{fees}{}
}
\value{

}
\description{
Cierra las ordenes pendientes al final del backtest
}
