% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{.start_indicator}
\alias{.start_indicator}
\title{Start of series indicator}
\usage{
.start_indicator(x)
}
\arguments{
\item{x}{xts}
}
\value{
returns 1 on the first observation
}
\description{
Start of series indicator
}
\examples{
.e
}
