% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{e_acf1}
\alias{e_acf1}
\title{Autocorrelation coefficient at lag 1 of the residuals}
\usage{
e_acf1(y)
}
\arguments{
\item{y}{a univariate time series}
}
\value{
A numeric value.
}
\description{
Computes the first order autocorrelation of the residual series of the deterministic trend model
}
\author{
Thiyanga Talagala
}
