% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{fast_recalc_stats}
\alias{fast_recalc_stats}
\title{Recalcula estadisticas sin analizar trades.
Obs. Solo recalcula un subconjunto de estadisticas}
\usage{
fast_recalc_stats(backtest_result, fees, quantity)
}
\arguments{
\item{backtest_result}{}

\item{fees}{}

\item{quantity}{}
}
\value{

}
\description{
Recalcula estadisticas sin analizar trades.
Obs. Solo recalcula un subconjunto de estadisticas
}
