% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fforms_ensemble.R
\name{fforms_ensemble}
\alias{fforms_ensemble}
\title{Function to identify models to compute combination forecast using FFORMS algorithm}
\usage{
fforms_ensemble(votematrix, threshold = 0.5)
}
\arguments{
\item{votematrix}{a matrix of votes of probabilities based of fforms random forest classifier}

\item{threshold}{threshold value for sum of probabilities of votes, default is 0.5}
}
\value{
a list containing the names of the forecast models
}
\description{
This function identify models to be use in producing combination forecast
}
\author{
Thiyanga Talagala
}
