% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fforms.R
\name{get_fforms}
\alias{get_fforms}
\title{Function to download pre-trained FFORMS random forests}
\usage{
get_fforms(frequency)
}
\arguments{
\item{frequency}{frequency of the data you will be forecasting, for example yearly, quarterly, monthly, weekly, daily and hourly}
}
\value{
an object class of randomForest
}
\description{
Load pretrained random forests trained based on FFORMS algorithm using the M1, M3 and M4 competition data
}
