% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTest.R
\name{head.ts}
\alias{head.ts}
\title{split a time series into training and test set}
\usage{
\method{head}{ts}(x, h)
}
\arguments{
\item{h}{length of the forecast horizon}

\item{y}{a univariate time series}
}
\value{
a univariate time series
}
\description{
the function head.ts extract the training set of a time series
}
\examples{
require(Mcomp)
quarterly_m3 <- subset(M3, "quarterly")
head.ts(quarterly_m3[[1]]$x, h=8)
}
\author{
Thiyanga Talagala
}
