% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{head_ts}
\alias{head_ts}
\title{split a time series into training and test set}
\usage{
head_ts(y, h)
}
\arguments{
\item{y}{a univariate time series}

\item{h}{length of the forecast horizon}
}
\value{
a univariate time series
}
\description{
the function head_ts extract the training set of a time series
}
\author{
Thiyanga Talagala
}
