% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{holtWinter_parameters}
\alias{holtWinter_parameters}
\title{Parameter estimates of Holt-Winters seasonal method}
\usage{
holtWinter_parameters(y)
}
\arguments{
\item{y}{a univariate time series}
}
\value{
A vector of 3 values: alpha, beta, gamma
}
\description{
Estimate the smoothing parameter for the level-alpha and
the smoothing parameter for the trend-beta, and seasonality-gamma
}
\author{
Thiyanga Talagala
}
