% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{is.bearish}
\alias{is.bearish}
\title{Comprueba si una vela es bajista.
Cierre < Apertura}
\usage{
is.bearish(df)
}
\arguments{
\item{df}{datos OHLC.}
}
\description{
Comprueba si una vela es bajista.
Cierre < Apertura
}
