% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{is.bullish}
\alias{is.bullish}
\title{Comprueba si una vela es alcista.
Cierre > Apertura}
\usage{
is.bullish(df)
}
\arguments{
\item{df}{datos OHLC.}
}
\description{
Comprueba si una vela es alcista.
Cierre > Apertura
}
