% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{is.counter_attack_line.bearish}
\alias{is.counter_attack_line.bearish}
\title{is.counter_attack_line.bearish}
\usage{
is.counter_attack_line.bearish(df)
}
\arguments{
\item{df}{datos OHLC.}
}
\description{
is.counter_attack_line.bearish
}
