% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{is.downtrend}
\alias{is.downtrend}
\title{Indica si es una tendencia bajista}
\usage{
is.downtrend(df, value = Cl, days = 5)
}
\arguments{
\item{df}{datos OHLC.}

\item{value}{el valor utilizado para el calculo.}

\item{days}{dias en consideracion.}
}
\value{
`TRUE` si es bajista, `FALSE` en otro caso.
}
\description{
Indica si es una tendencia bajista
}
