% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{is.separating_line.bullish}
\alias{is.separating_line.bullish}
\title{is.separating_line.bullish}
\usage{
is.separating_line.bullish(df)
}
\arguments{
\item{df}{datos OHLC.}
}
\description{
is.separating_line.bullish
}
