% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{is.three_methods.bearish}
\alias{is.three_methods.bearish}
\title{is.three_methods.bearish}
\usage{
is.three_methods.bearish(df)
}
\arguments{
\item{df}{datos OHLC.}
}
\description{
is.three_methods.bearish
}
