% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{is.three_methods.bullish}
\alias{is.three_methods.bullish}
\title{is.three_methods.bullish}
\usage{
is.three_methods.bullish(df)
}
\arguments{
\item{df}{datos OHLC.}
}
\description{
is.three_methods.bullish
}
