% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{is.three_white_soldiers}
\alias{is.three_white_soldiers}
\title{is.three_white_soldiers}
\usage{
is.three_white_soldiers(df)
}
\arguments{
\item{df}{datos OHLC.}
}
\description{
is.three_white_soldiers
}
