% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{is.uptrend}
\alias{is.uptrend}
\title{Indica si es una tendencia alcista.}
\usage{
is.uptrend(df, value = Cl, days = 5)
}
\arguments{
\item{df}{datos OHLC.}

\item{value}{el valor utilizado para el calculo.}

\item{days}{dias en consideracion.}
}
\value{
`TRUE` si es alcista, `FALSE` en otro caso.
}
\description{
Indica si es una tendencia alcista.
}
