% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{macd}
\alias{macd}
\title{MACD Strategy}
\usage{
macd(fastMA = 12, slowMA = 26, fees = 0, order_size = 1)
}
\arguments{
\item{fastMA}{periodos del fast MA}

\item{slowMA}{periodos del slow MA}

\item{fees}{costo numerico o funcion de costo}
}
\value{
un objeto Strategy
}
\description{
MACD Strategy
}
\examples{
strat <- macd()
}
