% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{macdhist}
\alias{macdhist}
\title{MACD hist Strategy}
\usage{
macdhist(fastMA = 12, slowMA = 26, fees = 0)
}
\arguments{
\item{fastMA}{periodos del fast MA}

\item{slowMA}{periodos del slow MA}

\item{fees}{costo numerico o funcion de costo}
}
\value{
un objeto Strategy
}
\description{
MACD hist Strategy
}
\examples{
strat <- macdhist()
}
