% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{Mcomp.sub}
\alias{Mcomp.sub}
\title{The following code is taken from Mcomp package and done some
small tweaks to handle hourly, daily and weekly data in M4 competition.}
\usage{
Mcomp.sub(x, getdata)
}
\arguments{
\item{x}{whether the data are from M1, M3 or M4}

\item{getdata}{condition: daily, monthly, demographic, etc.}
}
\description{
The following code is taken from Mcomp package and done some
small tweaks to handle hourly, daily and weekly data in M4 competition.
}
