% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlstrategy.R
\name{ml_apply_indicator_signals}
\alias{ml_apply_indicator_signals}
\title{Apply ML indicator and signals}
\usage{
ml_apply_indicator_signals(dataset, base_strategy, model, h, w)
}
\arguments{
\item{dataset}{the xts}

\item{base_strategy}{base technical analysis strategy}

\item{model}{the ML model}

\item{h}{prediction horizon}

\item{w}{historic window}
}
\value{

}
\description{
Apply ML indicator and signals
}
