% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlstrategy.R
\name{ml_strat}
\alias{ml_strat}
\title{ML strategy}
\usage{
ml_strat(base_strategy, approach, fees = 0, order_size = 1)
}
\arguments{
\item{base_strategy}{nombre de la estrategia de analisis tecnico}

\item{approach}{debe existir la funcion [approach]_signals. Por defecto puede ser una de `c("risky", "conservative")`}

\item{fees}{costo numerico o funcion de costo}

\item{order_size}{volumen de la orden}
}
\value{
un objeto Strategy
}
\description{
el dataset requiere las columnas
- enterLong, exitLong: Indicadores tecnicos
- pred_bull, pred_bear: Indicadores predictivos
}
\details{
Enter: prediction & signal
}
