% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{momentum}
\alias{momentum}
\title{Calcula el momento de una tendencia.}
\usage{
momentum(df, value = Cl, days = 5)
}
\arguments{
\item{df}{datos OHLC.}

\item{value}{el valor utilizado para el calculo.}

\item{days}{dias en consideracion.}
}
\value{
valor numérico.
}
\description{
Calcula el momento de una tendencia.
}
