% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlesticks.R
\name{multiple}
\alias{multiple}
\title{Indica posiciones que componen un patron de vela complejo.}
\usage{
multiple(flag, sessions = 2)
}
\arguments{
\item{flag}{vector logico que indica que posiciones son parte del patron de vela.}

\item{sessions}{numero de velas que componen el patron.}
}
\description{
Indica posiciones que componen un patron de vela complejo.
}
\examples{
multiple(flag = c(TRUE, TRUE, FALSE), sessions = 2)
}
