% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestplot.R
\name{old.plot_backtest}
\alias{old.plot_backtest}
\title{Returns a plot of the backtest}
\usage{
old.plot_backtest(strategy, dataset, backtest_results, range = "")
}
\arguments{
\item{strategy}{strategy object}

\item{dataset}{xts dataset}

\item{backtest_results}{backtest results from backtest.opt}
}
\value{
a ggplot
}
\description{
Returns a plot of the backtest
}
