% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{parse_params}
\alias{parse_params}
\title{Identifica parametros del experimento}
\usage{
parse_params(model_path)
}
\arguments{
\item{model_path}{path del archivo. Debe tener el formato "h[H]_w[W]_d[DATASET]_[METHOD].rds"}
}
\value{

}
\description{
Identifica parametros del experimento
}
