% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestplot.R
\name{plot_backtest_interactive}
\alias{plot_backtest_interactive}
\title{Creates an interactive plot}
\usage{
plot_backtest_interactive(...)
}
\arguments{
\item{...}{arguments for plot_backtest}
}
\value{
plotly plot object
}
\description{
Creates an interactive plot
}
