% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbacktestplot.R
\name{plotly_backtest}
\alias{plotly_backtest}
\title{Plot with interaction using plotly}
\usage{
plotly_backtest(...)
}
\arguments{
\item{...}{arguments for `plot_backtest`}
}
\value{

}
\description{
Plot with interaction using plotly
}
