% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict_models}
\alias{predict_models}
\title{meta-learning framework for forecast model selection}
\usage{
predict_models(meta_framework, ts_data_set)
}
\arguments{
\item{meta_framework}{classification framework that you want to use to identify forecast models}

\item{ts_data_set}{a list of univariate time series}
}
\value{
vector of best forecast models for all the series in the ts_data_set
}
\description{
meta-learning framework for forecast model selection
}
\author{
Thiyanga Talagala
}
