% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_xts}
\alias{predict_xts}
\title{Realiza prediccion utilizando un modelo de ML}
\usage{
predict_xts(model, x, h, w, overlap = TRUE)
}
\arguments{
\item{model}{modelo predictivo}

\item{x}{objeto xts}

\item{overlap}{TRUE para permitir predecir en cada observacion}
}
\value{
objeto xst con la columna de prediccion
}
\description{
Realiza prediccion utilizando un modelo de ML
}
