% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_xts_multiple}
\alias{predict_xts_multiple}
\title{Realiza prediccion utilizando un modelo de ML con multiples salidas}
\usage{
predict_xts_multiple(model, x, h, w, overlap = TRUE)
}
\arguments{
\item{model}{modelo predictivo}

\item{x}{objeto xts}

\item{overlap}{TRUE para permitir predecir en cada observacion}
}
\value{
objeto xts con las columnas de prediccion
}
\description{
Realiza prediccion utilizando un modelo de ML con multiples salidas
}
