% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_trainingset.R
\name{prepare_trainingset}
\alias{prepare_trainingset}
\title{preparation of training set}
\usage{
prepare_trainingset(accuracy_set, feature_set)
}
\arguments{
\item{accuracy_set}{output from the fcast_accuracy}

\item{feature_set}{output from the cal_features}
}
\value{
dataframe consisting features and classlabels
}
\description{
Preparation of a training set for random forest training
}
