% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{recalc_stats}
\alias{recalc_stats}
\title{Recalcula estadisticas con los nuevos parametros de compra/venta}
\usage{
recalc_stats(backtest_result, fees, quantity)
}
\arguments{
\item{backtest_result}{resultados de una ejecucion previa de backtest}

\item{fees}{costo de transaccion}

\item{quantity}{cantidad de compra/venta}
}
\value{

}
\description{
Recalcula estadisticas con los nuevos parametros de compra/venta
}
