% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{recalc_trades}
\alias{recalc_trades}
\title{Recalcula transacciones}
\usage{
recalc_trades(trades, fees = 0, qty = 1)
}
\arguments{
\item{trades}{datos de las transacciones}

\item{fees}{costo de transaccion}

\item{qty}{volumen de transaccion}
}
\value{
transacciones recalculadas
}
\description{
Recalcula transacciones
}
