% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtestplot.R
\name{recreate_strategy}
\alias{recreate_strategy}
\title{Recreate a strategy}
\usage{
recreate_strategy(backtest_results, fees = 0, order_size = 1)
}
\arguments{
\item{backtest_results}{backtest results list}

\item{fees}{cost per trade}

\item{order_size}{order volume}
}
\value{
strategy object
}
\description{
Recreate a strategy
}
