% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R, R/indicators.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_ohlcv}
\alias{EMA}
\alias{MACD}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{OHLCMerge}{\code{\link[OHLCMerge]{read_ohlcv}}}

  \item{TTR}{\code{\link[TTR]{EMA}}, \code{\link[TTR]{MACD}}}
}}

