% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfprediction_accuracy.R
\name{rfprediction_accuracy}
\alias{rfprediction_accuracy}
\title{function to calculate forecast-accuracy for new series}
\usage{
rfprediction_accuracy(predictions, tslist, database, function_name, h)
}
\arguments{
\item{predictions}{prediction results obtained from  random forest classifier}

\item{tslist}{list of new time series}

\item{database}{whethe the time series is from mcom or other}

\item{function_name}{to calculate accuracy measure, the arguments for the
accuracy function should be training period, test period and forecast}

\item{h}{length of the forecast horizon}
}
\value{
a numeric vector contains the forecast accuracy for each series
}
\description{
Given the prediction results of random forest calculate the
forecast-accuracy for the test set
}
