% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbacktest.R
\name{run_backtest}
\alias{run_backtest}
\title{Helper function to run the backtest}
\usage{
run_backtest(symbol, data, strat, cost = 0, qty = 1, ...)
}
\arguments{
\item{symbol}{name of the dataset or symbol}

\item{data}{OHLC data}

\item{strat}{strategy list object}

\item{cost}{cost for each trade}

\item{qty}{qty for the order}

\item{...}{other arguments for `backtest`}
}
\value{

}
\description{
Helper function to run the backtest
}
