% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer.R
\docType{package}
\name{seer-package}
\alias{seer}
\alias{seer-package}
\title{seer: functions for stock market prediction}
\description{
This package allows us to predict the direction of the stock market.
}
\author{
\strong{Maintainer}: Jordan Ayala \email{jmayalag@gmail.com}

}
