% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_arimabased.R
\name{sim_arimabased}
\alias{sim_arimabased}
\title{Simulate time series based on ARIMA models}
\usage{
sim_arimabased(y, Nsim, Combine = TRUE, M = TRUE)
}
\arguments{
\item{y}{a time series or M-competition data time series (Mcomp)}

\item{Nsim}{number of time series to simulate}

\item{Combine}{if TRUE, training and test data in the M-competition data are combined and generate a time
series corresponds to the full length of the series. Otherwise, it generate a time series
based on the training period of the series.}

\item{M}{if TRUE, y is considered to be a Mcomp data object}
}
\value{
A list of time series.
}
\description{
simulate multiple time series for a given series based on ARIMA models
}
\author{
Thiyanga Talagala
}
