% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlar.R
\name{stlar}
\alias{stlar}
\title{STL-AR method}
\usage{
stlar(y, h = 10, s.window = 11, robust = FALSE)
}
\arguments{
\item{y}{a univariate time series}

\item{h}{forecast horizon}

\item{s.window}{Either the character string “periodic” or the span (in lags) of the loess window for seasonal extraction}

\item{robust}{logical indicating if robust fitting be used in the loess procedue}
}
\value{
return object of class forecast
}
\description{
STL decomposition method applied to the time series, then an AR model is used to
forecast seasonally adjusted data, while the seasonal naive method is used to forecast
the seasonal component
}
\author{
Thiyanga Talagala
}
