% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.Mcomp}
\alias{subset.Mcomp}
\title{Subset of time series from the M Competitions
\code{subset.Mcomp} returns a subset of the time series data from the M
Competitions. Subsets can be for specific periods, or specific types of data
or both.}
\usage{
\method{subset}{Mcomp}(x, cond1, cond2, ...)
}
\arguments{
\item{x}{M-competition data or a subset of M-competition data}

\item{cond1}{Type or period of the data. Type is a character variable and
period could be character or numeric.}

\item{cond2}{Optional second condition specifying type or period of the
data, depending on \code{cond1}.  If \code{cond1} denotes type then
\code{cond2} would denote period, but if \code{cond1} denotes period then
\code{cond2} would denote type.}

\item{...}{Other arguments.}
}
\value{
An object of class \code{Mcomp} consisting of the selected series.
}
\description{
Possible values for \code{cond1} and \code{cond2} denoting period are 1, 4,
12, 24, 52, 365, "yearly", "quarterly", "monthly", "hourly", "weekly", "daily" and "other".
}
\details{
If \code{cond1} or \code{cond2} equals 111, then the 111 series used in the
extended comparisons in the 1982 M-competition are selected.

Possible values for \code{cond1} and \code{cond2} denoting type are "macro",
"micro", "industry", "finance", "demographic", "allother", "macro1",
"macro2", "micro1", "micro2", "micro3". These correspond to the descriptions
used in the competitions. See the references for details.

Partial matching used for both conditions.
}
\examples{
library(seer)
data(M4)
M4.quarterly <- subset(M4,4)
M4.yearly.industry <- subset(M4,1,"industry")
}
\references{
Rob Hyndman (2018). Mcomp: Data from the M-Competitions. R package version 2.7. https://CRAN.R-project.org/package=Mcomp
}
\author{
Thiyanga Talagala (Thiyanga has done small tweaks to adopt the code to M4data, original authors
of the code are Muhammad Akram and Rob Hyndman)
}
\keyword{data}
