% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTest.R
\name{tail.ts}
\alias{tail.ts}
\title{extract the test set of time series}
\usage{
\method{tail}{ts}(x, h)
}
\arguments{
\item{h}{length of the forecast horizon}

\item{y}{a univariate time series}
}
\value{
a univariate time series of length h
}
\description{
the function tail.ts extract the test set of a time series
}
\examples{
require(Mcomp)
quarterly_m3 <- subset(M3, "quarterly")
tail.ts(quarterly_m3[[1]]$x, h=8)
}
\author{
Thiyanga Talagala
}
