% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail.R
\name{tail_ts}
\alias{tail_ts}
\title{extract the test set of time series}
\usage{
tail_ts(y, h)
}
\arguments{
\item{y}{a univariate time series}

\item{h}{length of the forecast horizon}
}
\value{
a univariate time series of length h
}
\description{
the function tail_ts extract the test set of a time series
}
\author{
Thiyanga Talagala
}
