% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbacktest.R
\name{technical_ml}
\alias{technical_ml}
\title{Technical with ML strategy}
\usage{
technical_ml(base_strategy, model, h, w, approach = "risky", model_name = NULL)
}
\arguments{
\item{base_strategy}{the base technical strategy}

\item{model}{the ML model}

\item{h}{prediction horizon}

\item{w}{historical window}

\item{approach}{}

\item{model_name}{name of the model. If not specified, the class name will be used.}
}
\value{

}
\description{
Technical with ML strategy
}
