% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{time_series_prediction_format}
\alias{time_series_prediction_format}
\title{Convierte datos OHLC a formato requerido para prediccion de series de tiempo.}
\usage{
time_series_prediction_format(
  source,
  pred_col = "Close",
  max_horizon = 1,
  max_window = 2,
  discard = TRUE,
  overlap = FALSE
)
}
\arguments{
\item{source}{un OHLC de tipo xts}

\item{pred_col}{columna a utilizar en la prediccion}

\item{max_horizon}{maximo horizonte de prediccion}

\item{max_window}{maximo ventana a usar para el historico (w = 0 solo utiliza el dia actual )}

\item{discard}{descarta n primeros valores para que los restantes sean multiplos de w + h}

\item{overlap}{indica si se permite utilizar una observacion del tiempo t en mas de un caso.}
}
\value{
un data.frame formateado para la prediccion
}
\description{
Se puede especificar un valor alto de max_window y max_prediction_horizon si se va a reutilizar el data.frame
}
\examples{
\dontrun{
time_series_prediction_format(AAPL)
}
}
