% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{trade_stats}
\alias{trade_stats}
\title{Recalcula estadisticas en base a las transacciones}
\usage{
trade_stats(trades)
}
\arguments{
\item{trades}{xts con las txns del backtest}
}
\value{
data.frame con las estadisticas
}
\description{
Recalcula estadisticas en base a las transacciones
}
