% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_evtree}
\alias{train_evtree}
\title{Train evtree}
\usage{
train_evtree(training, fitControl, grid = NULL)
}
\arguments{
\item{training}{training dataset}

\item{fitControl}{parametros para el entrenamiento.}

\item{grid}{tuning grid}
}
\value{
the training information
}
\description{
Train evtree
}
