% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_nnet}
\alias{train_nnet}
\title{Entrena nnet
El data.frame solo debe tener las columnas de input, y el output con el nombre '.outcome'.}
\usage{
train_nnet(training, fitControl, grid)
}
\arguments{
\item{training}{data.frame de entrenamiento}

\item{fitControl}{parametros para el entrenamiento}

\item{grid}{espacio de busqueda de parametros}
}
\value{
resultados del entrenamiento
}
\description{
Entrena nnet
El data.frame solo debe tener las columnas de input, y el output con el nombre '.outcome'.
}
\examples{

\dontrun{
x <- time_series_prediction_format(AAPL, max_horizon=1, max_window=2)
train_nnet(x, grid=c(size=10, decay=0.1))
}
}
