% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_rf}
\alias{train_rf}
\title{Train Random Forest}
\usage{
train_rf(training, fitControl, grid = NULL)
}
\arguments{
\item{training}{training dataset}

\item{fitControl}{parametros para el entrenamiento.}

\item{grid}{tuning grid}
}
\value{
the training information
}
\description{
Train Random Forest
}
