% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trance.R
\name{trance}
\alias{trance}
\title{Enter a trance.}
\usage{
trance(data, type = c("main_effect", "interaction", "heatmap", "boxes_dots"))
}
\arguments{
\item{data}{Input data set to create graphs from}

\item{type}{'Trance' (visualization) type (eg. 'main_effect' for main effect results
for regression)}
}
\value{
Output the input dataframe with a new class added, to be used by
  \code{\link{visualize}}.
}
\description{
Start the visualization process by indicating the type of trance to enter,
ie. which type of visualization you would like.
}

