% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{triple_crossover}
\alias{triple_crossover}
\title{Triple Crossover Strategy
Consiste en tener tres MAs de diferentes velocidades
- Fast MA
- Medium MA
- Slow MA}
\usage{
triple_crossover(
  nFast = 10,
  nMedium = 25,
  nSlow = 50,
  fees = 0,
  order_size = 1
)
}
\arguments{
\item{nFast}{EMA Rapido}

\item{nMedium}{EMA Medio}

\item{nSlow}{EMA Lento}

\item{fees}{costo numerico o funcion de costo}

\item{order_size}{cantidad a comprar}
}
\value{
un objeto Strategy
}
\description{
Triple Crossover Strategy
Consiste en tener tres MAs de diferentes velocidades
- Fast MA
- Medium MA
- Slow MA
}
\examples{
strat <- triple_crossover()
}
