% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbacktest.R
\name{triple_ema}
\alias{triple_ema}
\title{Triple EMA strategy}
\usage{
triple_ema(nFast, nMedium, nSlow)
}
\arguments{
\item{nFast}{n for fast EMA}

\item{nMedium}{n for medium EMA}

\item{nSlow}{n for slow EMA}
}
\value{
a list object to pass to the backtest function
}
\description{
Triple EMA strategy
}
\examples{
triple_ema(1, 5, 10)
}
