% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{unitroot}
\alias{unitroot}
\title{Unit root test statistics}
\usage{
unitroot(y)
}
\arguments{
\item{y}{a univariate time series}
}
\value{
A vector of 3 values: test statistic based on PP-test and KPSS-test
}
\description{
Computes the test statistics based on unit root tests Phillips–Perron test and
KPSS test
}
\author{
Thiyanga Talagala
}
