% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_heatmap.R
\name{view_heatmap}
\alias{view_heatmap}
\title{Create a heatmap of values (usually correlation coefficients).}
\usage{
view_heatmap(data, y = names(data[1]), x = names(data[2]),
  value = names(data[3]), colours = c("skyblue4", "darkorange2"),
  number.colours = 2, values.text = TRUE, values.size = 5, ylab = "",
  xlab = "", legend.title = expression("Correlation" * ~rho))
}
\arguments{
\item{data}{The correlation dataset.}

\item{colours}{The spectrum of colours for the heatmap, as a vector between
the lowest (negative) value and the highest (positive) value}

\item{number.colours}{Number of colour spectrums to use.}

\item{values.size}{Font size of the values.}

\item{ylab}{Y-axis label.}

\item{xlab}{X-axis label.}

\item{legend.title}{Title of the legend.}

\item{values}{Whether or not the values are put on the plot.}
}
\value{
Creates a heatmap plot
}
\description{
Create a heatmap of values (usually correlation coefficients).
}
\examples{

\dontrun{
ds <- data.frame(Var1 = rep(letters[1:6], times = 6),
     Var2 = rep(letters[1:6], each = 6),
     Values = round(runif(36, min = -1, max = 1), 2))

view_heatmap(ds)
view_heatmap(ds, 'Var1', 'Var2', 'Values')
view_heatmap(ds, colours = c('darkblue', 'darkred'), values.text = FALSE)
view_heatmap(ds, 'Var1', 'Var2', 'Values', number.colours = 100)
view_heatmap(ds, 'Var1', 'Var2', 'Values', number.colours = 100, values.size = 10)
view_heatmap(ds, 'Var1', 'Var2', 'Values', number.colours = 100, values.size = 10)
}

}

