% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_pls_xloadings.R
\name{view_pls_xloadings}
\alias{view_pls_xloadings}
\title{Plot the x loadings by components for PLS models.}
\usage{
view_pls_xloadings(pls.model, comps = 1:2, renaming.x = function(x) x,
  title = NULL)
}
\arguments{
\item{pls.model}{The PLS model generated from \code{\link[pls]{plsr}}.}

\item{comps}{The component numbers, e.g. for the first 2 it would be '1:2',
for the first and third, 'c(1,3)', and so on.}

\item{renaming.x}{A function to renaming the x variables for the PLS model.}

\item{title}{}
}
\value{
Prints the loadings of the x in PLS by two components, with circles indicating
}
\description{
Plot the x loadings by components for PLS models.
}
\examples{

\dontrun{
library(pls)
data(yarn)

fit <- plsr(density ~ NIR, 6, data = yarn, validation = "CV")
view_pls_xloadings(fit)

fit <- plsr(density ~ NIR, 6, data = yarn)
view_pls_xloadings(fit)
view_pls_xloadings(fit, title = 'Near-infrared radiation')
view_pls_xloadings(fit, title = 'Near-infrared radiation',
 renaming.x = function(x) gsub('NIR', 'Wave ', x))
}
}

