% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_plsda_xloadings.R
\name{view_plsda_xloadings}
\alias{view_plsda_xloadings}
\title{Plot the x loadings by components for PLS-DA models.}
\usage{
view_plsda_xloadings(model, comps = 1:2, renaming.x = function(x) x,
  title = NULL)
}
\arguments{
\item{model}{The PLS-DA model generated from \code{\link[caret]{plsda}}.}

\item{comps}{The component numbers, e.g. for the first 2 it would be '1:2',
for the first and third, 'c(1,3)', and so on.}

\item{renaming.x}{A function to renaming the x variables for the PLS model.}

\item{title}{The main title of the plot.}
}
\value{
Prints the loadings of the X in PLS-DA by two components, with
 circles indicating 50% (dashed) and 100% (solid) explained variance.
}
\description{
Show the correlation circle plot of the loadings by component from the
partial least squares discriminant analysis.
}
\examples{

\dontrun{
# most of this example data is taken from the ?caret::plsda help.
library(caret)
data(mdrr)
set.seed(1)
inTrain <- sample(seq(along = mdrrClass), 450)
nzv <- nearZeroVar(mdrrDescr)
filteredDescr <- mdrrDescr[, -nzv]
training <- filteredDescr[inTrain,]
trainMDRR <- mdrrClass[inTrain]
preProcValues <- preProcess(training)
trainDescr <- predict(preProcValues, training)
fit   <- plsda(trainDescr, trainMDRR, ncomp = 5,
                    probMethod = "Bayes")
view_plsda_xloadings(fit)
view_plsda_xloadings(fit, title = 'Multidrug Resistance Reversal (MDRR) Agent Data: PLS-DA results')
view_plsda_xloadings(fit, title = 'Near-infrared radiation',
 renaming.x = function(x) gsub('RB', 'Protein ', x))
}
}

