% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision.R
\name{vision}
\alias{vision}
\alias{vision_simple}
\alias{vision_sparse}
\title{Vision experience (aka plot theme).}
\usage{
vision_sparse(base.plot, base.size = 12, base.family = "Arial")

vision_simple(base.plot, base.size = 12, base.family = "Arial",
  legend.position = c("right", "left", "bottom", "top", "none"))
}
\arguments{
\item{base_plot}{The base_plot sent from the \code{\link{visualize}} chain}

\item{base_size}{Font size}

\item{base_family}{Font family}
}
\value{
Customizes the appearance of the plot.
}
\description{
Describe the vision, also known as setting the theme for your plot.
}

