% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\alias{visualize.main_reg}
\title{Visualize the results.}
\usage{
visualize(data, ...)

\method{visualize}{main_reg}(data, groups = "~Yterms", ylab = "Exposures",
  xlab = "Beta estimates", center.line = c("0", "1", NULL))
}
\arguments{
\item{data}{Input data frame, sent from the \code{\link{trance}} command}

\item{...}{Additional parameters}

\item{groups}{Groups to facet the graph into (see
\code{\link[ggplot2]{facet_grid}}). Usage: 'horizontal~vertical',
'horizontal~.', or '~vertical'. So if a variable \code{Var} is set as
'Var~.' the variable will be set horizontally, while '~Var' will be set
vertically.}

\item{ylab}{Y-axis label}

\item{xlab}{X-axis label}

\item{center.line}{Where the dotted line is set}
}
\value{
Create a high level graphic.
}
\description{
After entering a trance, visualize the results of the analysis.
}

