% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{xts_to_df}
\alias{xts_to_df}
\title{Convierte un xts a un data.frame}
\usage{
xts_to_df(x)
}
\arguments{
\item{x}{el objeto xts}
}
\value{
un data.frame
}
\description{
Convierte un xts a un data.frame
}
