% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yearly_exp2}
\alias{yearly_exp2}
\title{Data related to experiment 2 - yearly series}
\format{A data frame with 1290645 rows and 26 variables
\describe{
\item{entropy}{spectral entropy}
\item{stability}{stability}
\item{hurst}{Hurst exponent}
\item{trend}{strength of trend}
\item{linearity}{linearity}
\item{y_acf1}{first autocorrelation coefficient of the original series}
\item{diff1y_acf1}{first autocorrelation coefficient of the differenced series}
\item{diff2y_acf1}{first autocorrelation coefficient of the twiced-differenced series}
\item{y_pacf5}{sum of squared of first 5 partial autoorrelation coefficient of the original series}
\item{diff1y_pacf5}{sum of squared of first 5 partial autocorrelation coefficients of the differenced series}
\item{diff2y_pacf5}{sum of squared of first 5 partial autocorrelation coefficients of the twiced-differenced series}
\item{beta}{Holt's linear trend model-beta hat}
\item{nonlinearity}{nonlinearity}
\item{N}{length of the time series}
\item{y_acf5}{sum of squared of first 5 autocorrelation coefficients of the original series}
\item{diff1y_acf5}{sum of squared of first 5 autocorrelation coefficients of the differenced series}
\item{diff2y_acf5}{sum of squared of first 5 autocorrelation coefficients of the twiced-differenced series}
\item{lmres_acf1}{first autocorelation coefficient of the residual series of linear trend model}
\item{ur_pp}{test statistics based on Phillips-Perron test}
\item{ur_kpss}{test statistics based on KPSS test}
\item{curvature}{curvature}
\item{spikiness}{spikiness}
\item{e_acf1}{first autocorrelation coefficient of the remainder series}
\item{lumpiness}{lumpiness}
\item{alpha}{Holt's linear trend model-alpha_hat}
\item{classlabel}{"best" forecasting method according to MASE}}}
\usage{
yearly_exp2
}
\description{
Data set related to Experiment 2 in yearly data
}
\examples{
data(yearly_exp2)
head(yearly_exp2)
summary(yearly_exp2)
}
\keyword{datasets}
