% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{acf5}
\alias{acf5}
\title{Autocorrelation-based features}
\usage{
acf5(x)
}
\arguments{
\item{x}{a univariate time series}
}
\value{
A vector of 3 values: sum of squared of first five autocorrelation coefficients of original series, first-differenced series,
and twice-differenced series.
}
\description{
Computes various measures based on autocorrelation coefficients of the
original series, first-differenced series and second-differenced series
}
\author{
Thiyanga Talagala
}
