% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_cal.R
\name{cal_features}
\alias{cal_features}
\title{Calculate features for new time series instances}
\usage{
cal_features(tslist, seasonal = FALSE, m = 1, lagmax = 2L, database,
  h, highfreq)
}
\arguments{
\item{tslist}{a list of univariate time series}

\item{seasonal}{if FALSE, restricts to features suitable for non-seasonal data}

\item{m}{frequency of the time series or minimum frequency in the case of msts objects}

\item{lagmax}{maximum lag at which to calculate the acf (quarterly series-5L, monthly-13L, weekly-53L, daily-8L, hourly-25L)}

\item{database}{whether the time series is from mcomp or other}

\item{h}{forecast horizon}

\item{highfreq}{whether the time series is weekly, daily or hourly}
}
\value{
dataframe: each column represent a feature and each row represent a time series
}
\description{
Computes relevant time series features before applying them to the model
}
\author{
Thiyanga Talagala
}
