% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fforms_combination.R
\name{fforms_combinationforecast}
\alias{fforms_combinationforecast}
\title{Combination forecast based on fforms}
\usage{
fforms_combinationforecast(fforms.ensemble, tslist, database, h,
  parallel = FALSE)
}
\arguments{
\item{fforms.ensemble}{a list output from fforms_ensemble function}

\item{tslist}{list of new time series}

\item{database}{whethe the time series is from mcom or other}

\item{h}{length of the forecast horizon}

\item{parallel}{If TRUE, multiple cores (or multiple sessions) will be used. This only speeds things up
when there are a large number of time series.}
}
\value{
a list containing, point forecast, confidence interval, accuracy measure
}
\description{
Compute combination forecast based on the vote matrix probabilities
}
\author{
Thiyanga Talagala
}
