% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbacktest.R
\name{macd}
\alias{macd}
\title{MACD strategy}
\usage{
macd(nFast, nSlow, nSig)
}
\arguments{
\item{nFast}{n for fast EMA}

\item{nSlow}{n for slow EMA}

\item{nSig}{n for signal EMA}
}
\value{
a list object to pass to the backtest function
}
\description{
MACD strategy
}
\examples{
macd(5, 12)
}
