% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbacktestplot.R
\name{plot_backtest}
\alias{plot_backtest}
\title{Plot backtest results}
\usage{
plot_backtest(result, range = "", additional_indicators)
}
\arguments{
\item{result}{backtest results}

\item{range}{xts range (e.g. "2019", "2018/2019")}
}
\value{

}
\description{
Plot backtest results
}
