% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer.R
\name{seer}
\alias{seer}
\title{SEER wrapper algorithm for ML method.}
\usage{
seer(y, X, learner = NULL, q0 = NULL, dmax = NULL, m = NULL,
  seed = 666, parallel_comput = T, nc = NULL)
}
\arguments{
\item{y}{A \code{vector} of binary response variable.}

\item{X}{A \code{matrix} or \code{data.frame} of attributes}

\item{q0}{A \code{numeric} value for the quantile of variable screening.}

\item{dmax}{A \code{double} representing the maximum number of attributes per learner.}

\item{m}{A \code{double} representing the maximum number learner per dimension explored}

\item{seed}{An \code{integer} that controls the reproducibility.}
}
\description{
SEER wrapper algorithm for ML method.
}
\examples{

}
\author{
Gaetan Bakalli and Samuel Orso
}
